setting.ymlのPriorityに小数を利用できるよう修正

+++ data/setting.go
@@ -3,11 +3,11 @@ package data
 type Setting struct {
 	BreakStr      string         `yaml:"BreakString"`
 	BreakPatterns []BreakPattern `yaml:"BreakPattern"`
-	PriorityWait  float32        `yaml:"PriorityWait"`
+	PriorityWait  float64        `yaml:"PriorityWait"`
 	DeleteFixs    []string       `yaml:"DeleteFix"`
 }
 
 type BreakPattern struct {
-	Priority int                 `yaml:"Priority"`
+	Priority float64             `yaml:"Priority"`
 	Patterns []map[string]string `yaml:"Pattern"`
 }
